+++ src/spawn/builtin/mod.rs
@@ -1,9 +1,11 @@
 //! Defines methods for spawning shell builtin commands
 
 mod colon;
+mod false_cmd;
 mod shift;
 mod true_cmd;
 
 pub use self::colon::{Colon, colon, SpawnedColon};
+pub use self::false_cmd::{False, false_cmd, SpawnedFalse};
 pub use self::shift::{Shift, shift, SpawnedShift};
 pub use self::true_cmd::{True, true_cmd, SpawnedTrue};